+++ mex/blueant/connector.py
@@ -24,7 +24,7 @@ class BlueAntConnector(HTTPConnector):
         api_key = settings.api_key.get_secret_value()
         self.session.headers["Authorization"] = f"Bearer {api_key}"
 
-    @cache  # noqa: B019
+    @cache
     def _get_json_from_api(self, relative_url: str) -> dict[str, Any]:
         """Get json from blueant api.