+++ mex/seq_repo/extract.py
@@ -1,6 +1,9 @@
 import json
 from collections.abc import Generator
 
+from mex.common.ldap.connector import LDAPConnector
+from mex.common.ldap.models.person import LDAPPersonWithQuery
+from mex.common.logging import watch
 from mex.seq_repo.model import (
     SeqRepoSource,
 )
@@ -22,3 +25,30 @@ def extract_sources() -> Generator[SeqRepoSource, None, None]:
         data = json.load(file)
         for item in data:
             yield SeqRepoSource.model_validate(item)
+
+
+@watch
+def extract_source_project_coordinator(
+    seq_repo_sources: dict[str, SeqRepoSource],
+) -> Generator[LDAPPersonWithQuery, None, None]:
+    """Extract LDAP persons with their query string for source project coordinators.
+
+    Args:
+        seq_repo_sources: Seq Repo sources
+
+    Returns:
+        Generator for LDAP persons with query
+    """
+    ldap = LDAPConnector.get()
+    seen = set()
+    for value in seq_repo_sources.values():
+        names = value.project_coordinators
+        for name in names:
+            if not name:
+                continue
+            if name in seen:
+                continue
+            seen.add(name)
+            persons = list(ldap.get_persons(mail=f"{name}@rki.de"))
+            if len(persons) == 1 and persons[0].objectGUID:
+                yield LDAPPersonWithQuery(person=persons[0], query=name)