+++ mex/seq_repo/transform.py
@@ -1,6 +1,7 @@
 from collections.abc import Generator
 from typing import Any
 
+from mex.common.ldap.models.person import LDAPPersonWithQuery
 from mex.common.models import (
     ExtractedAccessPlatform,
     ExtractedActivity,
@@ -8,13 +9,22 @@ from mex.common.models import (
     ExtractedPrimarySource,
     ExtractedResource,
 )
-from mex.common.types import Identifier
+from mex.common.types import (
+    Identifier,
+    MergedOrganizationalUnitIdentifier,
+    MergedPersonIdentifier,
+)
 from mex.seq_repo.model import SeqRepoSource
 
 
 def transform_seq_repo_activities_to_extracted_activities(
     seq_repo_sources: dict[str, SeqRepoSource],
     seq_repo_activity: dict[str, Any],
+    seq_repo_source_resolved_project_coordinators: list[LDAPPersonWithQuery],
+    unit_stable_target_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],
+    project_coordinators_merged_ids_by_query_string: dict[
+        str, list[MergedPersonIdentifier]
+    ],
     extracted_primary_source: ExtractedPrimarySource,
 ) -> Generator[ExtractedActivity, None, None]:
     """Transform seq-repo activity to ExtractedActivity.
@@ -22,23 +32,36 @@ def transform_seq_repo_activities_to_extracted_activities(
     Args:
         seq_repo_sources: Seq Repo extracted sources
         seq_repo_activity: Seq Repo extracted activity for default values from mapping
+        seq_repo_source_resolved_project_coordinators: Seq Repo sources resolved project
+                                            coordinators ldap query results
+        unit_stable_target_ids_by_synonym: Unit stable target ids by synonym
+        project_coordinators_merged_ids_by_query_string: Seq Repo Sources resolved
+                                                        project coordinators merged ids
         extracted_primary_source: Extracted primary source
 
     Returns:
         Generator for ExtractedActivity
     """
     theme = seq_repo_activity["theme"][0]["mappingRules"][0]["setValues"]
+
     for source in seq_repo_sources.values():
+        project_coordinators_ids, responsible_units = (
+            get_resolved_project_coordinators_and_units(
+                source.project_coordinators,
+                seq_repo_source_resolved_project_coordinators,
+                unit_stable_target_ids_by_synonym,
+                project_coordinators_merged_ids_by_query_string,
+            )
+        )
+
+        if not responsible_units or not project_coordinators_ids:
+            continue
         yield ExtractedActivity(
-            contact=[
-                Identifier.generate()
-            ],  # TODO: resolve 'source.project-coordinators' with ldap
+            contact=project_coordinators_ids,
             hadPrimarySource=extracted_primary_source.stableTargetId,
             identifierInPrimarySource=source.project_id,
-            involvedPerson=[],  # TODO: resolve 'source.project-coordinators' with ldap
-            responsibleUnit=[
-                Identifier.generate()
-            ],  # TODO: resolve 'source.project-coordinators' with ldap
+            involvedPerson=project_coordinators_ids,
+            responsibleUnit=responsible_units,
             theme=theme,
             title=source.project_name,
         )
@@ -87,6 +110,11 @@ def transform_seq_repo_resource_to_extracted_resource(
     seq_repo_distributions: dict[str, ExtractedDistribution],
     seq_repo_activities: dict[str, ExtractedActivity],
     seq_repo_resource: dict[str, Any],
+    seq_repo_source_resolved_project_coordinators: list[LDAPPersonWithQuery],
+    unit_stable_target_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],
+    project_coordinators_merged_ids_by_query_string: dict[
+        str, list[MergedPersonIdentifier]
+    ],
     extracted_primary_source: ExtractedPrimarySource,
 ) -> Generator[ExtractedResource, None, None]:
     """Transform seq-repo resource to ExtractedResource.
@@ -96,6 +124,11 @@ def transform_seq_repo_resource_to_extracted_resource(
         seq_repo_distributions: Seq Repo extracted distribution
         seq_repo_activities: Seq Repo extracted activity for default values from mapping
         seq_repo_resource: Seq Repo extracted resource
+        seq_repo_source_resolved_project_coordinators: Seq Repo sources resolved project
+                                            coordinators ldap query results
+        unit_stable_target_ids_by_synonym: Unit stable target ids by synonym
+        project_coordinators_merged_ids_by_query_string: Seq Repo Sources resolved
+                                                        project coordinators merged ids
         extracted_primary_source: Extracted primary source
 
     Returns:
@@ -126,16 +159,30 @@ def transform_seq_repo_resource_to_extracted_resource(
 
     for identifier_in_primary_source, source in seq_repo_sources.items():
         distribution = seq_repo_distributions[identifier_in_primary_source]
-        activity = seq_repo_activities[source.project_id]
+        activity = seq_repo_activities.get(source.project_id)
+
+        project_coordinators_ids, units_in_charge = (
+            get_resolved_project_coordinators_and_units(
+                source.project_coordinators,
+                seq_repo_source_resolved_project_coordinators,
+                unit_stable_target_ids_by_synonym,
+                project_coordinators_merged_ids_by_query_string,
+            )
+        )
+
+        if not units_in_charge or not project_coordinators_ids:
+            continue
+
+        contributing_unit = unit_stable_target_ids_by_synonym.get(
+            source.customer_org_unit_id
+        )
 
         yield ExtractedResource(
             accessRestriction=access_restriction,
             accrualPeriodicity=accrual_periodicity,
             anonymizationPseudonymization=anonymization_pseudonymization,
-            contact=[
-                Identifier.generate()
-            ],  # TODO: source.project_coordinators -> ldap
-            contributingUnit=[],  # TODO: source.customer_org_unit_id -> organigram
+            contact=project_coordinators_ids,
+            contributingUnit=contributing_unit or [],
             created=source.sequencing_date,
             distribution=distribution.stableTargetId,
             hadPrimarySource=extracted_primary_source.stableTargetId,
@@ -150,21 +197,21 @@ def transform_seq_repo_resource_to_extracted_resource(
             stateOfDataProcessing=state_of_data_processing,
             theme=theme,
             title=f"{source.project_name} sample {source.customer_sample_name}",
-            unitInCharge=[
-                Identifier.generate()
-            ],  # TODO: source.project_coordinators -> ldap, organigram
-            wasGeneratedBy=activity.stableTargetId,
+            unitInCharge=units_in_charge,
+            wasGeneratedBy=activity.stableTargetId if activity else None,
         )
 
 
 def transform_seq_repo_access_platform_to_extracted_access_platform(
     seq_repo_access_platform: dict[str, Any],
+    unit_stable_target_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],
     extracted_primary_source: ExtractedPrimarySource,
 ) -> ExtractedAccessPlatform:
     """Transform seq-repo access platform to ExtractedAccessPlatform.
 
     Args:
         seq_repo_access_platform: Seq Repo extracted access platform
+        unit_stable_target_ids_by_synonym: Unit stable target ids by synonym
         extracted_primary_source: Extracted primary source
 
     Returns:
@@ -192,9 +239,15 @@ def transform_seq_repo_access_platform_to_extracted_access_platform(
     ][0]["setValues"]
     title = seq_repo_access_platform["title"][0]["mappingRules"][0]["setValues"]
 
+    contacts = seq_repo_access_platform["contact"][0]["mappingRules"][0]["forValues"]
+
+    resolved_organigram = [
+        unit_stable_target_ids_by_synonym.get(contact) for contact in contacts
+    ]
+
     return ExtractedAccessPlatform(
         alternativeTitle=alternative_title,
-        contact=[],  # TODO: resolve with organigram
+        contact=resolved_organigram,
         description=description,
         endpointType=endpoint_type,
         hadPrimarySource=extracted_primary_source.stableTargetId,
@@ -202,5 +255,45 @@ def transform_seq_repo_access_platform_to_extracted_access_platform(
         landingPage=landing_page,
         technicalAccessibility=technical_accessibility,
         title=title,
-        unitInCharge=[],  # TODO: resolve with organigram
+        unitInCharge=resolved_organigram,
     )
+
+
+def get_resolved_project_coordinators_and_units(
+    project_coordinators: list[str],
+    seq_repo_source_resolved_project_coordinators: list[LDAPPersonWithQuery],
+    unit_stable_target_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],
+    project_coordinators_merged_ids_by_query_string: dict[
+        str, list[MergedPersonIdentifier]
+    ],
+) -> tuple[list[MergedPersonIdentifier], list[MergedOrganizationalUnitIdentifier]]:
+    """Get ldap resolved ids of project coordinators and units.
+
+    Args:
+        project_coordinators: Seq Repo raw project coordinator names
+        seq_repo_source_resolved_project_coordinators: Seq Repo sources resolved project
+                                            coordinators ldap query results
+        unit_stable_target_ids_by_synonym: Unit stable target ids by synonym
+        project_coordinators_merged_ids_by_query_string: Seq Repo Sources resolved
+                                                        project coordinators merged ids
+
+    Returns:
+        Resolved ids project coordinator and units
+    """
+    project_coordinators_ids = []
+    units_in_charge = []
+    for pc in project_coordinators:
+        person_merged_id = project_coordinators_merged_ids_by_query_string.get(pc)
+        if person_merged_id:
+            project_coordinators_ids.append(person_merged_id[0])
+
+        for query in seq_repo_source_resolved_project_coordinators:
+            query_ldap: LDAPPersonWithQuery = query
+            if (sam_account_name := query_ldap.person.sAMAccountName) and (
+                department_number := query_ldap.person.departmentNumber
+            ):
+                if sam_account_name.lower() == pc.lower():
+                    unit = unit_stable_target_ids_by_synonym.get(department_number)
+                    if unit and unit not in units_in_charge:
+                        units_in_charge.append(unit)
+    return project_coordinators_ids, units_in_charge